+++ Grid/Column/Column.php
@@ -83,6 +83,7 @@ abstract class Column
     protected $values = array();
     protected $selectFrom;
     protected $selectMulti;
+    protected $selectExpanded;
     protected $searchOnClick = false;
 
     protected $dataJunction = self::DATA_CONJUNCTION;
@@ -137,6 +138,7 @@ abstract class Column
         )));
         $this->setDefaultOperator($this->getParam('defaultOperator', self::OPERATOR_LIKE));
         $this->setSelectMulti($this->getParam('selectMulti', false));
+        $this->setSelectExpanded($this->getParam('selectExpanded', false));
         $this->setSearchOnClick($this->getParam('searchOnClick'), false);
     }
 
@@ -737,6 +739,16 @@ abstract class Column
         $this->selectMulti = $selectMulti;
     }
 
+    public function getSelectExpanded()
+    {
+        return $this->selectExpanded;
+    }
+
+    public function setSelectExpanded($selectExpanded)
+    {
+        $this->selectExpanded = $selectExpanded;
+    }
+
     public function hasDQLFunction(&$matches = null)
     {
         $regex = '/.(?P<all>(?P<field>\w+):(?P<function>\w+)(:)?(?P<parameters>\w*))$/';
@@ -774,8 +786,7 @@ abstract class Column
      */
     public function isFilterSubmitOnChange()
     {
-        if ($this->getSelectMulti()) return false;
-        return true;
+        return !$this->getSelectMulti();
     }
 
     public function setSearchOnClick($searchOnClick)